Reworked quickstart to use atoms
