+++ website/src/components/CodeEditor.astro
@@ -1,5 +1,5 @@
 ---
-import { getSandpackCssText } from "@codesandbox/sandpack-react";
+import { getSandpackCssText, type SandpackSetup } from "@codesandbox/sandpack-react";
 import { CodeEditor as ReactCodeEditor } from "./CodeEditor";
 
 import type { ComponentProps } from "react";
@@ -8,16 +8,17 @@ type SandpackProps = ComponentProps<typeof ReactCodeEditor>;
 type Props = {
   template?: SandpackProps["template"];
   files?: SandpackProps["files"];
+  dependencies?: SandpackSetup["dependencies"];
 };
 
 const styles = getSandpackCssText();
-const { template, files } = Astro.props;
+const { template, files, dependencies } = Astro.props;
 
 // See docs: https://sandpack.codesandbox.io/docs/getting-started/usage#templates
 ---
 <style id="sandpack" is:global set:html={styles} />
 
-<!-- This wrapper ensures the default Atro Starlight theme doesn't add stupid marings
+<!-- This wrapper ensures the default Atro Starlight theme doesn't add stupid margins
 to all the elements in Sandpack's UI -->
 <div class="not-content">
   <ReactCodeEditor
@@ -25,6 +26,7 @@ to all the elements in Sandpack's UI -->
     client:load
     customSetup={{
       dependencies: {
+        ...dependencies,
         // Must exist
         bunshi: "latest",
       },
