+++ website/src/source-examples/quickstart/App.vue
@@ -1,22 +1,7 @@
 <script setup>
-import { useMolecule } from "bunshi/vue";
-import { ref, onMounted, onUnmounted } from "vue";
-import { ResizeMolecule } from "./molecules";
-
-const observer = useMolecule(ResizeMolecule);
-const root = ref(null);
-onMounted(() => observer.observe(root.value));
-onUnmounted(() => observer.unobserve(root.value));
+import Counter from "./Counter.vue";
 </script>
-<style>
-.resize {
-  border: 2px solid;
-  padding: 20px;
-  width: 300px;
-  resize: both;
-  overflow: auto;
-}
-</style>
 <template>
-  <div ref="root" class="resize">My size is being watched.</div>
+  <Counter />
+  <Counter />
 </template>
