+++ website/src/source-examples/quickstart/Counter.vue
@@ -1,0 +1,15 @@
+<script setup>
+import { useMolecule } from "bunshi/vue";
+import { CountMolecule } from "./molecules";
+import { useAtom } from "jotai-vue";
+
+const countAtom = useMolecule(CountMolecule);
+const [count, setCount] = useAtom(countAtom);
+</script>
+
+<template>
+  <div>
+    <p>Times clicked: {{ count }}</p>
+    <button @click="setCount((prev) => prev + 1)">increment</button>
+  </div>
+</template>
