+++ website/src/source-examples/quickstart/molecules2.ts
@@ -1,0 +1,7 @@
+import { ComponentScope, molecule } from "bunshi";
+import { atom } from "jotai/vanilla";
+
+export const CountMolecule = molecule((_, scope) => {
+  scope(ComponentScope);
+  return atom(0);
+});
