+++ website/src/source-examples/quickstart/sources.ts
@@ -1,11 +1,15 @@
 import ReactApp from "./App.tsx?raw";
 import VueApp from "./App.vue?raw";
+import VueCounter from "./Counter.vue?raw";
 import molecules from "./molecules.ts?raw";
+import molecules2 from "./molecules2.ts?raw";
 
 const quickstart = {
-    molecules,
-    VueApp,
-    ReactApp,
+  molecules,
+  molecules2,
+  VueApp,
+  VueCounter,
+  ReactApp,
 };
 
 export default quickstart;
